Allow ssh to host node via commctl.
With the addition of the HostCreds resource it's now possible explicitly
gain access to the remote_user and ssh_priv_key for a host. This patch
requests the ssh private key and user, creates a temporary file to house
the key, and then uses that key and remote user with ssh to create a
connection. After the connection is finished (or if the connection fails
to be made) the local copy of the ssh key is deleted.